Removed net mode selection
With the final engine desync being fixed in packet-server mode, peer-to-peer mode has little reason to be used and will be unable to support certain features in the future. As such, packet-server will now always be how net games are handled for better stability and a more expected experience alongside being easier to maintain overall.